#9755: centralize data movement C++ implementation to the data_movement/concat folder
